modified:   app/Http/Controllers/Api/CartController.php 	modified:   app/Http/Controllers/Api/Management/ProductController.php 	modified:   resources/views/profile/partials/update-profile-information-form.blade.php

+++ app/Http/Controllers/Api/Management/ProductController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -30,6 +31,12 @@ class ProductController extends Controller
      */
     public function store(Product $product, Request $request)
     {
+        if (!ProductController::auth()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Youhave to be login first'
+            ]);
+        }
         // Validation
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255',
@@ -37,15 +44,15 @@ class ProductController extends Controller
             'quantity' => 'required|numeric|integer',
             'image' => 'required|image|mimes:jpg,png,gif,jpeg,svg|max:2048',
             'price' => 'required|decimal:0,5'
-            ]);
-            if ($validator->fails()) {
-                return response()->json([
+        ]);
+        if ($validator->fails()) {
+            return response()->json([
                 'success' => false,
                 'message' => 'Product creating failed',
                 'error' => $validator->messages()
             ], 201);
         }
-        
+
         // Fetch the product data
         $product = $request->all();
         $product['price'] = round($product['price'], 5);
@@ -88,6 +95,12 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+        if (!ProductController::auth()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'You have to be login first'
+            ]);
+        }
         // Validation 
         $validator = Validator::make($request->all(), [
             'name' => 'string|max:255',
@@ -105,7 +118,10 @@ class ProductController extends Controller
             ], 201);
         }
         $newProduct = $request->all();
-        $newProduct['price'] = round($newProduct['price'], 5);
+        
+        if (isset($newProduct['price'])) {
+            $newProduct['price'] = round($newProduct['price'], 5);
+        }
 
         $product = Product::find($product->id);
 
@@ -141,6 +157,13 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        if(!ProductController::auth()){
+            return response()->json([
+                'success' => false,
+                'message' => 'You have to be login first'
+            ]);
+        }
+
         if ($product) {
             $product->delete();
             return response()->json([
@@ -152,5 +175,12 @@ class ProductController extends Controller
             'success' => false,
             'message' => 'Product not found',
         ], 404);
+    }
+
+    public function auth(){
+        if(Auth::guest()){
+            return false;
+        }
+        return true;
     }
 }
